Cast location to u16 in rival trainer switch
Explicitly casts show->rivalTrainer.location to u16 in the switch statement to ensure correct type handling and prevent potential issues with implicit type conversion.

+++ src/tv.c
@@ -5519,7 +5519,7 @@ static void DoTVShowTodaysRivalTrainer(void)
     switch (state)
     {
     case 0:
-        switch (show->rivalTrainer.location)
+        switch ((u16)show->rivalTrainer.location)
         {
         default:
             sTVShowState = 7;
